Udemy-44  Using state in function components
We can use function components as well,
but there is different approach to manage the state.

We manage it by invoking setState funtion, which returns two object array.
First property is actually the state that has been set,
and second one is a funtion that should be used to update this state.

We may set multiple 0-level states to avoid overriding in case of updating.