Udemy-47 Two way binding
With seeting value on input to props.name we make it
display value from property 'name' in it as typed.
Without onChange this bahaves odd as it doesn't allow us to change input state
(our input is beeing constantly overrided by static value of property).

In onChange we can set binding to listener method handling changes of state.